Fix up some logging, add server-side logging for messages from frontend
