+++ lib/chooseUserGrid.js
@@ -8,6 +8,22 @@ const promptForUserGridChoice = () => {
   });
 };
 
+const recieveUserGridChoice = async (store) => {
+  let promptObj = {
+    type: "prompt",
+    prompt: "Choose a grid number:",
+  };
+  store.ws.send(JSON.stringify(promptObj));
+  return new Promise((resolve) => {
+    store.ws.on("message", (message) => {
+      let gridObj = JSON.parse(message);
+      if (gridObj.type === "grid") {
+        return resolve(gridObj.grid);
+      }
+    });
+  });
+};
+
 const isValidUserGridChoice = (store, choice) => {
   if (
     isNaN(choice) ||
@@ -30,7 +46,9 @@ const getUserGridChoice = async (store) => {
     store.log(`${player}'s turn:`);
   }
   while (true) {
-    let grid = await promptForUserGridChoice();
+    let grid = store.ws
+      ? await recieveUserGridChoice(store)
+      : await promptForUserGridChoice();
     if (isValidUserGridChoice(store, grid) === true) {
       const log = store.singlePlayer
         ? `You chose ${grid}!`
