+++ lib/play.js
@@ -7,26 +7,27 @@ const { pickOrder, setPlayOrder } = require("./playOrder");
 const { playAgain } = require("./playAgain");
 const Store = require("./store");
 
-const recieveUserGridChoice = async (store) => {
-  let promptObj = {
-    type: "prompt",
-    prompt: "Choose a grid number:",
-  };
-  store.ws.send(JSON.stringify(promptObj));
-  return new Promise((resolve) => {
-    store.ws.on("message", (message) => {
-      let gridObj = JSON.parse(message);
-      if (gridObj.type === "grid") {
-        return resolve(gridObj.grid);
-      }
-    });
-  });
-};
+// const recieveUserGridChoice = async (store) => {
+//   let promptObj = {
+//     type: "prompt",
+//     prompt: "Choose a grid number:",
+//   };
+//   store.ws.send(JSON.stringify(promptObj));
+//   return new Promise((resolve) => {
+//     store.ws.on("message", (message) => {
+//       let gridObj = JSON.parse(message);
+//       if (gridObj.type === "grid") {
+//         return resolve(gridObj.grid);
+//       }
+//     });
+//   });
+// };
 
 const userTurn = async (store) => {
-  const userChoice = store.ws
-    ? await recieveUserGridChoice(store)
-    : await getUserGridChoice(store);
+  // const userChoice = store.ws
+  //   ? await recieveUserGridChoice(store)
+  //   : await getUserGridChoice(store);
+  const userChoice = await getUserGridChoice(store);
   store.setUserGridChoice(userChoice);
 };
